implementing lazy loading for cold start (#33)
* implementing lazy loading for cold start

* Fixing linter errors

* Removing excess firebase-functions to improve performance in case of cold start

* Fixing linter

* Continuing to minimize imports
