+++ functions/controllers/announcements-api/index.js
@@ -1,10 +1,11 @@
 const functions = require("firebase-functions");
-const helmet = require("helmet");
 const express = require("express");
 const cors = require("cors");
-const { corsConfig } = require("../../utils/config");
 const { jwtCheck, validKey, hasDeleteAnnouncement, hasUpdateAnnouncement } = require("../../utils/middleware");
 const { addDocument, queryCollectionSorted, deleteDocument } = require("../../utils/database");
+
+const auth0Config = functions.config().auth;
+const corsConfig = auth0Config ? auth0Config.cors : "";
 
 const announcements = express();
 const corsOptions = {
@@ -13,7 +14,6 @@ const corsOptions = {
 };
 
 announcements.disable("x-powered-by");
-announcements.use(helmet());
 announcements.use(cors(corsOptions));
 announcements.use(express.json());
 
@@ -84,4 +84,6 @@ announcements.post("/", jwtCheck, hasUpdateAnnouncement, async (req, res) => {
     });
 });
 
-module.exports = { announcements };
+const service = functions.https.onRequest(announcements);
+
+module.exports = { announcements, service };
